refactor: ConcurrentMemoryRepository
